Move states setup into `setup()`
It really should've been there in the first place. Now it not having been present in the mod when installed wont cause issues.

It'll cause it to build it as normal, but if it has issues with the root element of the same name still exists, that's a migration issue, not our fault.